change all short to int

+++ main.cpp
@@ -2,13 +2,13 @@
 #include <vector>
 using std::vector;
 
-vector<short> fibonacci(short x, short y, short it) {
-  short temp = x;
+vector<int> fibonacci(int x, int y, int it) {
+  int temp = x;
   x = y;
   y += temp;
   it = it + 1;
   std::cout << it << ": " << x << "\n";
-  vector<short> result;
+  vector<int> result;
   result[1] = x;
   result[2] = y;
   result[3] = it;
@@ -16,9 +16,9 @@ vector<short> fibonacci(short x, short y, short it) {
 }
 
 int main() {
-  short it = 1; short x = 1; short y = 1;
+  int it = 1; int x = 1; int y = 1;
   while (it < 15) {
-    vector<short> result = fibonacci(x, y, it);
+    vector<int> result = fibonacci(x, y, it);
     x = result[1]; y = result[2]; it = result[3];
     }
   return 0;
